Update to double precision and add README.md

+++ src/stat_utils.rs
@@ -4,6 +4,6 @@ use crate::dawg::Dawg;
 use rusty_dawg::stat_utils;
 
 #[pyfunction]
-pub fn good_turing_estimate(dawg: &Dawg, n_tokens: usize) -> f32 {
+pub fn good_turing_estimate(dawg: &Dawg, n_tokens: usize) -> f64 {
     stat_utils::good_turing_estimate(dawg.get_dawg(), n_tokens)
 }
